Fix timezone issue in integration test
Allow setting a fixed timezone in the plugin settings
so the modification list timestamps are as expected.

The timezone plugin setting is deliberately not exposed via plugin UI
at the moment because it is already quite overloaded and generally using
the system default timezone seems sufficient for our use cases.